requirements.txt
Flask==3.1.2
torch==2.1.0
torchvision==0.16.0
opencv-python==4.8.1.78
numpy==1.24.3
Werkzeug==3.1.3
Pillow==10.1.0
gunicorn==21.2.0
python-dotenv==1.0.0
